Allow sub-day reminder intervals for easier push testing
The interval dropdown gains "10 min", "30 min", and "1 hour" options
on top of the existing day-scale choices. Picking a sub-day interval
hides the notify-time picker, since wall-clock anchoring doesn't make
sense at that frequency.

- Backend: intervalDays is now Float (was Int) in the GraphQL schema,
  with min validation relaxed from 1 day to ~1 minute. The existing
  computeNextDueAt math already worked for fractions.
- Frontend: intervalOptions becomes an array of {days, labelKey, n}
  so we can mix minute/hour/day labels. round6 helper keeps float
  values stable across the ngModel round-trip.
- i18n: minuteCount and hourCount labels in en + da.

Test recipe: set water reminder to 10 min, wait ~10 min, open the app
to trigger a GraphQL request — push fires within seconds.

+++ backend/src/schema.ts
@@ -164,7 +164,7 @@ export const typeDefs = gql`
     id: String!
     plantId: String!
     actionType: ReminderActionType!
-    intervalDays: Int!
+    intervalDays: Float!
     notifyTime: String
     nextDueAt: String!
     snoozedUntil: String
@@ -229,7 +229,7 @@ export const typeDefs = gql`
     adminCreateUser(email: String!, password: String!, role: String!, tier: String!): User!
     adminUpdateUser(userId: String!, email: String, role: String, tier: String): User!
     adminDeleteUser(userId: String!): Boolean!
-    setPlantReminder(plantId: String!, actionType: ReminderActionType!, intervalDays: Int!, enabled: Boolean!, notifyTime: String): PlantReminder!
+    setPlantReminder(plantId: String!, actionType: ReminderActionType!, intervalDays: Float!, enabled: Boolean!, notifyTime: String): PlantReminder!
     removePlantReminder(id: String!): Boolean!
     snoozeReminder(id: String!, hours: Int): PlantReminder!
     subscribeToPush(subscription: PushSubscriptionInput!): Boolean!
